requirements.txt
Django==4.2.7
Pillow==10.1.0
django-crispy-forms==2.1
crispy-bootstrap5==0.7
python-decouple==3.8
openpyxl==3.1.2
reportlab==4.2.0
djangorestframework==3.15.1
drf-yasg==1.21.7 